chore: remove unused DirectionMapper, move DirectionCharacteristic to common/samm
