+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/supply/logic/dto/daysofsupplysamm/DaysOfSupply.java
@@ -23,7 +23,7 @@ import java.util.HashSet;
 import java.util.Objects;
 
 import org.eclipse.tractusx.puris.backend.common.util.PatternStore;
-import org.eclipse.tractusx.puris.backend.stock.logic.dto.itemstocksamm.DirectionCharacteristic;
+import org.eclipse.tractusx.puris.backend.common.samm.DirectionCharacteristic;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
